Removed types.ts for Curves
Moved all the type declarations into curves_webgl to prevent circular imports
