+++ src/bundles/curve/curves_webgl.ts
@@ -1,6 +1,7 @@
+/* eslint-disable max-classes-per-file */
 /* eslint-disable @typescript-eslint/naming-convention */
 import { mat4, vec3 } from 'gl-matrix';
-import { ProgramInfo, BufferInfo, Point } from './types';
+import { ModuleState } from 'js-slang';
 
 /** @hidden */
 export const drawnCurves: CurveDrawn[] = [];
@@ -83,8 +84,42 @@ function initShaderProgram(
   return shaderProgram;
 }
 
+/** Encapsulates 3D point with RGB values. */
+export type Point = {
+  x: number;
+  y: number;
+  z: number;
+  color: [r: number, g: number, b: number, t: number];
+};
+
+/** A function that takes in CurveFunction and returns a tranformed CurveFunction. */
+export type CurveTransformer = (c: Curve) => Curve;
+
+/**
+ * The return type of the curve generated by the source code in workspace.
+ * WebGLCanvas in Curves tab captures the return value and calls its init function.
+ */
+type ProgramInfo = {
+  program: WebGLProgram;
+  attribLocations: {
+    vertexPosition: number;
+    vertexColor: number;
+  };
+  uniformLocations: {
+    projectionMatrix: WebGLUniformLocation | null;
+    modelViewMatrix: WebGLUniformLocation | null;
+  };
+};
+
+type BufferInfo = {
+  cubeBuffer: WebGLBuffer | null;
+  curveBuffer: WebGLBuffer | null;
+  curveColorBuffer: WebGLBuffer | null;
+};
+
 /** A function that takes in number from 0 to 1 and returns a Point. */
 export type Curve = (t: number) => Point;
+
 /**
  * A function that specifies additional rendering information when taking in
  * a CurveFunction and returns a ShapeDrawn based on its specifications.
@@ -127,7 +162,6 @@ export class CurveDrawn {
     let max_y = -Infinity;
     let min_z = Infinity;
     let max_z = -Infinity;
-
 
     for (let i = 0; i <= numPoints; i += 1) {
       const point = func(i / numPoints);
@@ -406,4 +440,12 @@ export class CurveDrawn {
       gl.drawArrays(gl.POINTS, 0, this.numPoints + 1);
     }
   };
+}
+
+export class CurveModuleState implements ModuleState {
+  constructor() {
+    this.drawnCurves = [];
+  }
+
+  public drawnCurves: CurveDrawn[];
 }
